LARSplitting2D.crossingTest — Function

Half-line crossing test. Utility function for pointInPolygonClassification function. Update the count depending of the actual crossing of the tile half-line.

LARSplitting2D.setTile — Function

setTile(box)(point)

Set the tileCode of the 2D bbox [b1,b2,b3,b4]:=[ymax,ymin,xmax,xmin]:= x,x,y,y including the 2D point of x,y coordinates. Depending on point position, tileCode ranges in 0:15, and uses bit operators. Used to set the plane tiling depending on position of the query point, in order to subsequently test the tile codes of edges of a 2D polygon, and determine if the query point is either internal, external, or on the boundary of the polygon. Function to be parallelized ...

c1,c2 = tilecode(p1),tilecode(p2)
c_edge, c_un, c_int = c1 ⊻ c2, c1 | c2, c1 & c2

LARSplitting2D.input_collection — Function

input_collection(data::Array)::Tuple

Facet selection. Construction of a $(d-1)$-dimensional collection from a $(d-1)$- or $d$-dimensional one. $0-chain$ of LAR type are used as input.

Output is $admissible input$ for algorithms of the 2D/3D arrangement pipeline.

Example 2D

An assembly of geometric objects is generated, and their assembly, including rotated and translated chains, is built producing a collection of input LAR models.

V,(_,EV,FV) = Lar.cuboidGrid([4,4],true);
W,(_,EW,FW) = Lar.cuboidGrid([3,5],true);
mycircle(r,n) = Lar.circle(r)(n)

data2d1 = (V,EV)
data2d2 = Lar.Struct([ Lar.t(2,2), Lar.r(pi/3), Lar.t(-1.5,-2.5), (W,EW) ])
data2d3 = Lar.Struct([ Lar.t(2,2), mycircle(2.5,16) ])
data2d4 = Lar.Struct([ Lar.t(3.5,3.5), mycircle(.25,16) ])
data2d5 = Lar.Struct([ Lar.t(5,3.5), mycircle(.5,16) ])
data2d6 = Lar.Struct([ Lar.t(5,3.5), mycircle(.25,16) ])

model2d = input_collection( [ data2d1, data2d2, data2d3, data2d4, data2d5, data2d6 ] )
V,EV = model2d
VV = [[k] for k in 1:size(V,2)];
using Plasm
Plasm.view( Plasm.numbering(.5)((V,[VV,EV])) )

Note that V,EV is not a cellular complex, since 1-cells intersect out of 0-cells.

Example 3D

V,FV = Lar.sphere(2)([3,4])
EV = Lar.simplexFacets(FV)
mysphere = V,FV,EV

data3d1 = mysphere
data3d2 = Lar.Struct([ Lar.t(0,1,0), mysphere ])
data3d3 = Lar.Struct([ Lar.t(0,0.5,0), Lar.s(0.4,0.4,0.4), mysphere ])
data3d4 = Lar.Struct([ Lar.t(4,0,0), Lar.s(0.8,0.8,0.8), mysphere ])
data3d5 = Lar.Struct([ Lar.t(4,0,0), Lar.s(0.4,0.4,0.4), mysphere ])

model3d = input_collection([ data3d1, data3d2, data3d3, data3d4, data3d5 ])
V,FV,EV = model3d
VV = [[k] for k in 1:size(V,2)];
using Plasm
Plasm.view( Plasm.numbering(1)((V,[VV, EV])) )

Note that V,FV,EV is not a cellular complex, since 1-cells and 2-cells intersect out of 0-cells.

Missing docstring.

Missing docstring for LARSplitting2D.boundingbox. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LARSplitting2D.coordintervals. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LARSplitting2D.boxcovering. Check Documenter's build log for details.

LARSplitting2D.spaceindex — Function

spaceindex(model::Lar.LAR)::Array{Array{Int,1},1}

Generation of space indexes for all $(d-1)$-dim cell members of model.

Spatial index made by $d$ interval-trees on bounding boxes of $sigma in S_{d−1}$. Spatial queries solved by intersection of $d$ queries on IntervalTrees generated by bounding-boxes of geometric objects (LAR cells).

The return value is an array of arrays of ints, indexing cells whose containment boxes are intersecting the containment box of the first cell. According to Hoffmann, Hopcroft, and Karasick (1989) the worst-case complexity of Boolean ops on such complexes equates the total sum of such numbers.

Examples 2D

julia> V = hcat([[0.,0],[1,0],[1,1],[0,1],[2,1]]...);

julia> EV = [[1,2],[2,3],[3,4],[4,1],[1,5]];

julia> Sigma = Lar.spaceindex((V,EV))
5-element Array{Array{Int64,1},1}:
 [4, 5, 2]
 [1, 3, 5]
 [4, 5, 2]
 [1, 3, 5]
 [4, 1, 3, 2]

From model2d value, available in ?input_collection docstring:

julia> Sigma =  spaceindex(model2d);

Example 3D

model = model3d
Sigma =  spaceindex(model3d);
Sigma

LARSplitting2D.intersection — Function

intersection(line1,line2)

Intersect two line segments in 2D, by computing the two line parameters of the intersection point.

The line segments intersect if both return parameters α,β are contained in the interval [0,1].

Example

julia> line1 = [[0.,0], [1,2]]
2-element Array{Array{Float64,1},1}:
 [0.0, 0.0]
 [1.0, 2.0]

julia> line2 = [[2.,0], [0,3]]
2-element Array{Array{Float64,1},1}:
 [2.0, 0.0]
 [0.0, 3.0]

julia> Lar.intersection(line1,line2)
(0.8571428571428571, 0.5714285714285714)

LARSplitting2D.linefragments — Function

linefragments(V,EV,Sigma)

Compute the sequences of ordered parameters fragmenting each input lines.

Extreme parameter values (0.0 and 1.0) are included in each output line. Sigma is the spatial index providing the subset of lines whose containment boxes intersect the box of each input line (given by EV).

julia> V = hcat([[0.,0],[1,0],[1,1],[0,1],[2,1]]...);

julia> EV = [[1,2],[2,3],[3,4],[4,1],[1,5]];

julia> Sigma = Lar.spaceindex((V,EV))
5-element Array{Array{Int64,1},1}:
 [4, 5, 2]
 [1, 3, 5]
 [4, 5, 2]
 [1, 3, 5]
 [4, 1, 3, 2]

julia> Lar.linefragments(V,EV,Sigma)
5-element Array{Any,1}:
 [0.0, 1.0]
 [0.0, 0.5, 1.0]
 [0.0, 1.0]
 [0.0, 1.0]
 [0.0, 0.5, 1.0]

LARSplitting2D.fragmentlines — Function

fragmentlines(model::Lar.LAR)::Lar.LAR

Pairwise intersection of 2D line segments.

Example 2D

V,EV = model2d
W, EW = Lar.fragmentlines(model2d) # OK
using Plasm
Plasm.viewexploded(W,EW)(1.2,1.2,1.2)

Missing docstring.

Missing docstring for LARSplitting2D.fraglines. Check Documenter's build log for details.

LARSplitting2D.congruence — Function

congruence(model::Lar.LAR)::Lar.LAR

Graded bases of equivalence classes Ck (Uk ), with Uk = Xk /Rk for 0 ≤ k ≤ 2.

Example

julia>